parsing_llm/requirements.txt
requests
docx2txt
striprtf
python-dotenv
fastapi
python-multipart
uvicorn